finish splitting the test files
